重构一部分httpclient
