+++ webmagic-core/src/main/java/us/codecraft/webmagic/downloader/HttpClientDownloader.java
@@ -3,16 +3,16 @@ package us.codecraft.webmagic.downloader;
 import org.apache.commons.io.IOUtils;
 import org.apache.http.HttpHost;
 import org.apache.http.HttpResponse;
-import org.apache.http.NameValuePair;
 import org.apache.http.annotation.ThreadSafe;
-import org.apache.http.client.config.CookieSpecs;
-import org.apache.http.client.config.RequestConfig;
-import org.apache.http.client.entity.UrlEncodedFormEntity;
+import org.apache.http.auth.AuthState;
+import org.apache.http.auth.UsernamePasswordCredentials;
 import org.apache.http.client.methods.CloseableHttpResponse;
 import org.apache.http.client.methods.HttpUriRequest;
-import org.apache.http.client.methods.RequestBuilder;
+import org.apache.http.client.protocol.HttpClientContext;
+import org.apache.http.impl.auth.BasicScheme;
 import org.apache.http.impl.client.CloseableHttpClient;
-import org.apache.http.message.BasicNameValuePair;
+import org.apache.http.protocol.BasicHttpContext;
+import org.apache.http.protocol.HttpContext;
 import org.apache.http.util.EntityUtils;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
@@ -23,12 +23,13 @@ import us.codecraft.webmagic.Task;
 import us.codecraft.webmagic.proxy.Proxy;
 import us.codecraft.webmagic.selector.PlainText;
 import us.codecraft.webmagic.utils.CharsetUtils;
-import us.codecraft.webmagic.utils.HttpConstant;
 import us.codecraft.webmagic.utils.WMCollections;
 
 import java.io.IOException;
 import java.nio.charset.Charset;
-import java.util.*;
+import java.util.HashMap;
+import java.util.Map;
+import java.util.Set;
 
 
 /**
@@ -46,9 +47,15 @@ public class HttpClientDownloader extends AbstractDownloader {
 
     private HttpClientGenerator httpClientGenerator = new HttpClientGenerator();
 
-    private CloseableHttpClient getHttpClient(Site site, Proxy proxy) {
+    private HttpUriRequestConverter httpUriRequestConverter = new HttpUriRequestConverter();
+
+    public void setHttpUriRequestConverter(HttpUriRequestConverter httpUriRequestConverter) {
+        this.httpUriRequestConverter = httpUriRequestConverter;
+    }
+
+    private CloseableHttpClient getHttpClient(Site site) {
         if (site == null) {
-            return httpClientGenerator.getClient(null, proxy);
+            return httpClientGenerator.getClient(null);
         }
         String domain = site.getDomain();
         CloseableHttpClient httpClient = httpClients.get(domain);
@@ -56,7 +63,7 @@ public class HttpClientDownloader extends AbstractDownloader {
             synchronized (this) {
                 httpClient = httpClients.get(domain);
                 if (httpClient == null) {
-                    httpClient = httpClientGenerator.getClient(site, proxy);
+                    httpClient = httpClientGenerator.getClient(site);
                     httpClients.put(domain, httpClient);
                 }
             }
@@ -66,35 +73,31 @@ public class HttpClientDownloader extends AbstractDownloader {
 
     @Override
     public Page download(Request request, Task task) {
-        Site site = null;
-        if (task != null) {
-            site = task.getSite();
+        if (task == null || task.getSite() == null) {
+            throw new NullPointerException("task or site can not be null");
         }
-        Set<Integer> acceptStatCode;
-        String charset = null;
-        Map<String, String> headers = null;
-        if (site != null) {
-            acceptStatCode = site.getAcceptStatCode();
-            charset = site.getCharset();
-            headers = site.getHeaders();
-        } else {
-            acceptStatCode = WMCollections.newHashSet(200);
-        }
-        logger.info("downloading page {}", request.getUrl());
+        logger.debug("downloading page {}", request.getUrl());
         CloseableHttpResponse httpResponse = null;
         int statusCode = 0;
+        Site site = task.getSite();
         try {
-            HttpHost proxyHost = null;
-            Proxy proxy = null; //TODO
-            if (site != null && site.getHttpProxyPool() != null && site.getHttpProxyPool().isEnable()) {
+            Proxy proxy = null;
+            if (site.getHttpProxyPool() != null && site.getHttpProxyPool().isEnable()) {
                 proxy = site.getHttpProxyFromPool();
-                proxyHost = proxy.getHttpHost();
             } else if (site != null && site.getHttpProxy() != null){
-                proxyHost = site.getHttpProxy();
+                proxy = site.getHttpProxy();
+                request.putExtra(Request.PROXY, site.getHttpProxy());
             }
-            
-            HttpUriRequest httpUriRequest = getHttpUriRequest(request, site, headers, proxyHost);
-            httpResponse = getHttpClient(site, proxy).execute(httpUriRequest);
+            request.putExtra(Request.PROXY, proxy);
+
+            HttpContext httpContext = new BasicHttpContext();
+
+            HttpUriRequest httpUriRequest = httpUriRequestConverter.convert(request,site);
+            AuthState authState = new AuthState();
+            authState.update(new BasicScheme(), new UsernamePasswordCredentials("userName", "password"));
+            httpContext.setAttribute(HttpClientContext.PROXY_AUTH_STATE, authState);
+            CloseableHttpClient httpClient = getHttpClient(site, proxy);
+            httpResponse = httpClient.execute(httpUriRequest, httpContext);
             statusCode = httpResponse.getStatusLine().getStatusCode();
             request.putExtra(Request.STATUS_CODE, statusCode);
             if (statusAccept(acceptStatCode, statusCode)) {
@@ -132,72 +135,6 @@ public class HttpClientDownloader extends AbstractDownloader {
 
     protected boolean statusAccept(Set<Integer> acceptStatCode, int statusCode) {
         return acceptStatCode.contains(statusCode);
-    }
-
-    protected HttpUriRequest getHttpUriRequest(Request request, Site site, Map<String, String> headers, HttpHost proxy) {
-        RequestBuilder requestBuilder = selectRequestMethod(request).setUri(request.getUrl());
-        if (headers != null) {
-            for (Map.Entry<String, String> headerEntry : headers.entrySet()) {
-                requestBuilder.addHeader(headerEntry.getKey(), headerEntry.getValue());
-            }
-        }
-
-        RequestConfig.Builder requestConfigBuilder = RequestConfig.custom();
-        if (site != null) {
-            requestConfigBuilder.setConnectionRequestTimeout(site.getTimeOut())
-                    .setSocketTimeout(site.getTimeOut())
-                    .setConnectTimeout(site.getTimeOut())
-                    .setCookieSpec(CookieSpecs.BEST_MATCH);
-        }
-
-        if (proxy != null) {
-			requestConfigBuilder.setProxy(proxy);
-			request.putExtra(Request.PROXY, proxy);
-		}
-        requestBuilder.setConfig(requestConfigBuilder.build());
-        return requestBuilder.build();
-    }
-
-    protected RequestBuilder selectRequestMethod(Request request) {
-        String method = request.getMethod();
-        if (method == null || method.equalsIgnoreCase(HttpConstant.Method.GET)) {
-            //default get
-            return addQueryParams(RequestBuilder.get(),request.getParams());
-        } else if (method.equalsIgnoreCase(HttpConstant.Method.POST)) {
-            return addFormParams(RequestBuilder.post(), (NameValuePair[]) request.getExtra("nameValuePair"), request.getParams());
-        } else if (method.equalsIgnoreCase(HttpConstant.Method.HEAD)) {
-            return addQueryParams(RequestBuilder.head(),request.getParams());
-        } else if (method.equalsIgnoreCase(HttpConstant.Method.PUT)) {
-            return addFormParams(RequestBuilder.put(), (NameValuePair[]) request.getExtra("nameValuePair"), request.getParams());
-        } else if (method.equalsIgnoreCase(HttpConstant.Method.DELETE)) {
-            return addQueryParams(RequestBuilder.delete(),request.getParams());
-        } else if (method.equalsIgnoreCase(HttpConstant.Method.TRACE)) {
-            return addQueryParams(RequestBuilder.trace(),request.getParams());
-        }
-        throw new IllegalArgumentException("Illegal HTTP Method " + method);
-    }
-
-    private RequestBuilder addFormParams(RequestBuilder requestBuilder, NameValuePair[] nameValuePair, Map<String, String> params) {
-        List<NameValuePair> allNameValuePair=new ArrayList<NameValuePair>();
-        if (nameValuePair != null && nameValuePair.length > 0) {
-            allNameValuePair= Arrays.asList(nameValuePair);
-        }
-        if (params != null) {
-            for (String key : params.keySet()) {
-                allNameValuePair.add(new BasicNameValuePair(key, params.get(key)));
-            }
-        }
-        requestBuilder.setEntity(new UrlEncodedFormEntity(allNameValuePair, Charset.forName("utf8")));
-        return requestBuilder;
-    }
-
-    private RequestBuilder addQueryParams(RequestBuilder requestBuilder, Map<String, String> params) {
-        if (params != null) {
-            for (Map.Entry<String, String> entry : params.entrySet()) {
-                requestBuilder.addParameter(entry.getKey(), entry.getValue());
-            }
-        }
-        return requestBuilder;
     }
 
     protected Page handleResponse(Request request, String charset, HttpResponse httpResponse, Task task) throws IOException {
