+++ webmagic-core/src/main/java/us/codecraft/webmagic/downloader/HttpClientGenerator.java
@@ -1,13 +1,9 @@
 package us.codecraft.webmagic.downloader;
 
-import org.apache.commons.lang3.StringUtils;
 import org.apache.http.HttpException;
 import org.apache.http.HttpRequest;
 import org.apache.http.HttpRequestInterceptor;
-import org.apache.http.auth.AuthScope;
-import org.apache.http.auth.UsernamePasswordCredentials;
 import org.apache.http.client.CookieStore;
-import org.apache.http.client.CredentialsProvider;
 import org.apache.http.config.Registry;
 import org.apache.http.config.RegistryBuilder;
 import org.apache.http.config.SocketConfig;
@@ -21,7 +17,6 @@ import org.apache.http.protocol.HttpContext;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import us.codecraft.webmagic.Site;
-import us.codecraft.webmagic.proxy.Proxy;
 
 import javax.net.ssl.SSLContext;
 import javax.net.ssl.TrustManager;
@@ -92,30 +87,12 @@ public class HttpClientGenerator {
         return this;
     }
 
-    public CloseableHttpClient getClient(Site site, Proxy proxy) {
-        return generateClient(site, proxy);
+    public CloseableHttpClient getClient(Site site) {
+        return generateClient(site);
     }
 
-    private CloseableHttpClient generateClient(Site site, Proxy proxy) {
-        CredentialsProvider credsProvider = null;
+    private CloseableHttpClient generateClient(Site site) {
         HttpClientBuilder httpClientBuilder = HttpClients.custom();
-        
-        if (proxy != null && StringUtils.isNotBlank(proxy.getUser()) && StringUtils.isNotBlank(proxy.getPassword()))
-        {
-            credsProvider= new BasicCredentialsProvider();
-            credsProvider.setCredentials(
-                    new AuthScope(proxy.getHttpHost().getAddress().getHostAddress(), proxy.getHttpHost().getPort()),
-                    new UsernamePasswordCredentials(proxy.getUser(), proxy.getPassword()));
-            httpClientBuilder.setDefaultCredentialsProvider(credsProvider);
-        }
-
-        if (site != null && site.getHttpProxy()!= null && site.getUsernamePasswordCredentials() != null){
-            credsProvider = new BasicCredentialsProvider();
-            credsProvider.setCredentials(
-                    new AuthScope(site.getHttpProxy()),//可以访问的范围
-                    site.getUsernamePasswordCredentials());//用户名和密码
-            httpClientBuilder.setDefaultCredentialsProvider(credsProvider);
-        }
         
         httpClientBuilder.setConnectionManager(connectionManager);
         if (site != null && site.getUserAgent() != null) {
