+++ webmagic-core/src/main/java/us/codecraft/webmagic/proxy/ProxyPool.java
@@ -6,7 +6,10 @@ import org.apache.http.HttpHost;
  * Created by edwardsbean on 15-2-28.
  */
 public interface ProxyPool {
-    public void returnProxy(HttpHost host, int statusCode);
-    public Proxy getProxy();
-    public boolean isEnable();
+
+    void returnProxy(HttpHost host, int statusCode);
+
+    Proxy getProxy();
+    
+    boolean isEnable();
 }
