+++ webmagic-core/src/main/java/us/codecraft/webmagic/proxy/TimerReuseProxy.java
@@ -1,0 +1,163 @@
+package us.codecraft.webmagic.proxy;
+
+import java.io.Serializable;
+import java.util.ArrayList;
+import java.util.List;
+import java.util.concurrent.Delayed;
+import java.util.concurrent.TimeUnit;
+
+/**    
+ * >>>> Proxy lifecycle 
+ 
+        +----------+     +-----+
+        | last use |     | new |
+        +-----+----+     +---+-+
+              |  +------+   |
+              +->| init |<--+
+                 +--+---+
+                    |
+                    v
+                +--------+
+           +--->| borrow |
+           |    +---+----+
+           |        |+------------------+
+           |        v
+           |    +--------+
+           |    | in use |  Respone Time
+           |    +---+----+
+           |        |+------------------+
+           |        v
+           |    +--------+
+           |    | return |
+           |    +---+----+
+           |        |+-------------------+
+           |        v
+           |    +-------+   reuse interval
+           |    | delay |   (delay time)
+           |    +---+---+
+           |        |+-------------------+
+           |        v
+           |    +------+
+           |    | idle |    idle time
+           |    +---+--+
+           |        |+-------------------+
+           +--------+
+ */
+
+/**
+ * Object has these status of lifecycle above.<br>
+ * 
+ * @author yxssfxwzy@sina.com <br>
+ * @since 0.5.1
+ * @see TimerReuseProxyPool
+ */
+
+public class TimerReuseProxy extends Proxy implements Delayed, Serializable {
+
+	private static final long serialVersionUID = 228939737383625551L;
+	public static final int ERROR_403 = 403;
+	public static final int ERROR_404 = 404;
+	public static final int ERROR_BANNED = 10000;// banned by website
+	public static final int ERROR_Proxy = 10001;// the proxy itself failed
+	public static final int SUCCESS = 200;
+
+	private int reuseTimeInterval = 1500;// ms
+	private Long canReuseTime = 0L;
+	private Long lastBorrowTime = System.currentTimeMillis();
+	private Long responseTime = 0L;
+
+	private int failedNum = 0;
+	private int successNum = 0;
+	private int borrowNum = 0;
+
+	private List<Integer> failedErrorType = new ArrayList<Integer>();
+
+	public TimerReuseProxy(ProxyHost proxyHost, String user, String password) {
+		super(proxyHost, user, password);
+	}
+
+	public TimerReuseProxy(ProxyHost proxyHost, String user, String password, int reuseTimeInterval) {
+		super(proxyHost, user, password);
+		this.reuseTimeInterval = reuseTimeInterval;
+	}
+
+	public int getSuccessNum() {
+		return successNum;
+	}
+
+	public void successNumIncrement(int increment) {
+		this.successNum += increment;
+	}
+
+	public Long getLastUseTime() {
+		return lastBorrowTime;
+	}
+
+	public void setLastBorrowTime(Long lastBorrowTime) {
+		this.lastBorrowTime = lastBorrowTime;
+	}
+
+	public void recordResponse() {
+		this.responseTime = (System.currentTimeMillis() - lastBorrowTime + responseTime) / 2;
+		this.lastBorrowTime = System.currentTimeMillis();
+	}
+
+	public List<Integer> getFailedErrorType() {
+		return failedErrorType;
+	}
+
+	public void setFailedErrorType(List<Integer> failedErrorType) {
+		this.failedErrorType = failedErrorType;
+	}
+
+	public void fail(int failedErrorType) {
+		this.failedNum++;
+		this.failedErrorType.add(failedErrorType);
+	}
+
+	public void setFailedNum(int failedNum) {
+		this.failedNum = failedNum;
+	}
+
+	public int getFailedNum() {
+		return failedNum;
+	}
+
+	public String getFailedType() {
+		String re = "";
+		for (Integer i : this.failedErrorType) {
+			re += i + " . ";
+		}
+		return re;
+	}
+
+	public int getReuseTimeInterval() {
+		return reuseTimeInterval;
+	}
+
+	public void setReuseTimeInterval(int reuseTimeInterval) {
+		this.reuseTimeInterval = reuseTimeInterval;
+		this.canReuseTime = System.nanoTime() + TimeUnit.NANOSECONDS.convert(reuseTimeInterval, TimeUnit.MILLISECONDS);
+
+	}
+
+	@Override
+	public long getDelay(TimeUnit unit) {
+		return unit.convert(canReuseTime - System.nanoTime(), TimeUnit.NANOSECONDS);
+	}
+
+	@Override
+	public int compareTo(Delayed o) {
+		TimerReuseProxy that = (TimerReuseProxy) o;
+		return canReuseTime > that.canReuseTime ? 1 : (canReuseTime < that.canReuseTime ? -1 : 0);
+
+	}
+
+	public void borrowNumIncrement(int increment) {
+		this.borrowNum += increment;
+	}
+
+	public int getBorrowNum() {
+		return borrowNum;
+	}
+}
